extract bisko dataclasses
